move getting Refs into getFirestoreData, remove try catch
