+++ src/utils/index.js
@@ -3,6 +3,5 @@ export { default as setupLogic } from './setupLogic/index';
 export * from './componentUtils';
 export * from './currentDoc';
 export * from './getFirestoreData';
-export * from './getFirestoreRefs';
 export * from './setupLogic/index';
 export * from './dummyState';
